requirements.txt
google-cloud-texttospeech>=2.15.0 